adding isUnique constraint on field type
When building forms, it could be useful to limit the number of instances of a field type on a form.

This property, `isUnique`, permits this behavior.

On the toolbar, when attempting to add a field type that has already been used, if the isUnique property is set, the builder will deny adding the field.